+++ CHANGES.rst
@@ -6,6 +6,7 @@ grafana-wtf changelog
 in progress
 ===========
 - Add proof-of-concept implementation
+- Add Grafana API key token authentication
 
 
 2019-01-24 0.0.0